+++ mpc_energy/config_manager.py
@@ -15,9 +15,18 @@ accepted_risks = get_entity_id("accepted_risks")
 MQTT_USER = get_entity_id("ha_mqtt_user")
 MQTT_PASS = get_entity_id("ha_mqtt_pass")
 
-amber_api_key = get_entity_id("amber_api_key")
+energy_retailer = get_entity_id("energy_retailer", default="amber")
+demand_price = get_entity_id("demand_price", default="")
+demand_window_start = get_entity_id("demand_window_start", default="16:00")
+demand_window_end = get_entity_id("demand_window_end", default="21:00")
+
+amber_api_key = get_entity_id("amber_api_key", default="")
 amber_site_id = get_entity_id("amber_site_id", default="")
-amber_demand_price = get_entity_id("amber_demand_price", default="")
+
+flow_import_price_entity_id = get_entity_id("flow_import_price_entity_id", default="")
+flow_export_price_entity_id = get_entity_id("flow_export_price_entity_id", default="")
+flow_price_forecast_entity_id = get_entity_id("flow_price_forecast_entity_id", default="")
+
 battery_discharge_cost = get_entity_id("battery_discharge_cost")
 spike_price_warning_level = get_entity_id("spike_price_warning_level", default=25)
 notification_target = get_entity_id("notification_target", default="")